Stop crash in 1.12.5 when drawing elytra (#204)
Game would crash when rendering a player wearing elytra. `isWearingChestplate()` would return true, renderer crashes upon trying to draw chestplate model. This is a quick-and-dirty fix to avoid this specific crash. If a player is wearing a chestplate which happens to have a GLIDER component, the chestplate will not be rendered.